added dragonfly; using mixin to draw md5 entities; some entity classes now abstract; added Weap enum

+++ src/ants/rescache.d
@@ -5,7 +5,7 @@ mixin template ResourceCacheMixin(T, KT=string)
   union ResourcePtr
   {
     Resource r;
-    int p;
+    size_t p;
   }
 
   struct ResourcePionter
